\name{prop_names}
\alias{prop_names}
\docType{data}
\title{
Names of aminoacid properties
}
\description{
Names of all properties in \code{aa_prop}.
}

\details{
A list of two elements. First element \code{long} contains long names and second 
\code{short} short names
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(prop_names)
## maybe str(prop_names) ; plot(prop_names) ...
}
\keyword{datasets}
